fix: replace vote counting with similarity-weighted scoring in voyage key selection
COUNT(*) biases towards longer voyages (more chunks = more votes regardless of
relevance). SUM(1 - distance) per voyage rewards voyages whose chunks are
actually close to the query embedding, reducing false winners from long voyages
with many mediocre chunks.

+++ log/log_voyage_key_testing.py
@@ -15,7 +15,7 @@ def log_voyage_key_testing(
     returned_keys: list[str],
     hit: bool,
     winner_rank: int | None,
-    vote_counts: dict[str, int],
+    vote_counts: dict[str, float],
 ) -> None:
     """Logs a single question result to the test_voyage_key_logging table."""
     with conn.cursor() as cur:
